Improved NDJSON
