+++ examples/json_read.rs
@@ -3,23 +3,17 @@ use std::io::BufReader;
 use std::sync::Arc;
 
 use arrow2::array::Array;
-use arrow2::error::{ArrowError, Result};
+use arrow2::error::Result;
 use arrow2::io::json::read;
 
 fn read_path(path: &str) -> Result<Arc<dyn Array>> {
     // Example of reading a JSON file.
     let reader = BufReader::new(File::open(path)?);
-    let data = serde_json::from_reader(reader)?;
+    let json = serde_json::from_reader(reader)?;
 
-    let values = if let serde_json::Value::Array(values) = data {
-        Ok(values)
-    } else {
-        Err(ArrowError::InvalidArgumentError("".to_string()))
-    }?;
+    let data_type = read::infer(&json)?;
 
-    let data_type = read::infer_rows(&values)?;
-
-    Ok(read::deserialize_json(&values, data_type))
+    read::deserialize(&json, data_type)
 }
 
 fn main() -> Result<()> {
