+++ src/io/ndjson/mod.rs
@@ -1,2 +1,3 @@
-//! APIs to read NDJSON
+//! APIs to read from and write to NDJSON
 pub mod read;
+pub mod write;
